fix(highlight): link HLF_8 to something more prominent
Problem:
HLF_8 is used by :intro and :map. And none of the other HLF_x things
use the other Special groups. So we have a darkgray highlight that's
very common, and not easy to read.

Darkgray makes sense for "Conceal", "non-printing chars", etc. But not
for keycodes or similar common elements.

Solution:
Link SpecialKey to Special.